Refactoring code to SOLID architecture
This update was extremely crucial, weren't added new features, the code only was refactored to be prepared to grow and scale.

Was implemented DDD Domain Driven Design architecture.

Folders was moved and was created basicly two main layers called modules and shared.

Inside modules was created two folders to separete responsabilities, to deal with appointments and to deal with users.

Inside each module was created a infra folder to deal with infra layer, that was separated to the root module folder.

Inside shared was created a container folder to deal with injection dependency and infra folder to deal with infraescructure shared.

Was created too controllers folder inside http folder of infra, to deal with logical of routes.